Refactor WS API into hmda-platform project
