+++ hmda/src/main/scala/hmda/api/http/HmdaWSApi.scala
@@ -1,13 +1,12 @@
 package hmda.api.http
 
 import akka.actor.{ActorSystem, Props}
-import akka.pattern.pipe
 import akka.event.Logging
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.server.Route
+import akka.pattern.pipe
 import akka.stream.ActorMaterializer
 import com.typesafe.config.ConfigFactory
-import hmda.api.http.model.common.HttpServer
 
 import scala.concurrent.{ExecutionContext, Future}
